Вынес интеграционный тест планировщика
